refactor user schemas files
